+++ R/get_pd_vi_scores.R
@@ -5,38 +5,45 @@
 #'
 #' @param object A fitted model object (e.g., a \code{"randomForest"} object).
 #'
-#' @param pred.var Character string giving the names of the predictor variables
-#' of interest.
+#' @param feature_names Character string giving the names of the predictor
+#' variables of interest.
 #'
 #' @param FUN Function used to measure the variability of the partial dependence
 #' values for continuous predictors. If \code{NULL}, the standard deviation is
 #' used (i.e., \code{FUN = sd}). For factors, the range statistic is used (i.e.,
 #' (max - min) / 4).
 #'
-#' @param keep.partial Logical indicating whether or not to return the computed
-#' partial dependence values for each predictor listed in \code{pred.var} in a
-#' separate attribute called \code{"partial"}. Default is \code{FALSE}.
+#' @param keep_partial Logical indicating whether or not to return the computed
+#' partial dependence values for each predictor listed in \code{feature_names}
+#' in a separate attribute called \code{"partial"}. Default is \code{FALSE}.
 #'
 #' @param ... Additional optional arguments to be passed onto
 #' \code{\link{get_pd_vi_score}}.
 #'
 #' @keywords internal
-get_pd_vi_scores <- function(object, pred.var, FUN = NULL, keep.partial = FALSE,
-                             ...) {
-  imp <- lapply(pred.var, function(x) {
-    get_pd_vi_score(object, pred.var = x, FUN = FUN,
-                    keep.partial = keep.partial, ...)
+get_pd_vi_scores <- function(object, feature_names, FUN = NULL,
+                             keep_partial = FALSE, ...) {
+
+  # Print warning message
+  warning("Using `partial = TRUE` is experimental, use at your own risk!",
+          call. = FALSE)
+
+  # Compute partial dependence-based variable importance scores
+  importance_scores <- lapply(feature_names, function(x) {
+    get_pd_vi_score(object, feature_names = x, FUN = FUN,
+                    keep_partial = keep_partial, ...)
   })
-  names(imp) <- pred.var
-  if (!is.null(attr(imp[[1L]], "partial"))) {
-    pd <- lapply(imp, FUN = function(x) attr(x, "partial"))
-    names(pd) <- pred.var
-    imp <- unlist(imp)
-    attr(imp, "partial") <- pd
-    imp
+  names(importance_scores) <- feature_names
+  if (!is.null(attr(importance_scores[[1L]], "partial"))) {
+    pd <- lapply(importance_scores, FUN = function(x) attr(x, "partial"))
+    names(pd) <- feature_names
+    importance_scores <- unlist(importance_scores)
+    attr(importance_scores, "partial") <- pd
+    importance_scores
   } else {
-    unlist(imp)
+    unlist(importance_scores)
   }
+
 }
 
 
@@ -47,38 +54,47 @@ get_pd_vi_scores <- function(object, pred.var, FUN = NULL, keep.partial = FALSE,
 #'
 #' @param object A fitted model object (e.g., a \code{"randomForest"} object).
 #'
-#' @param pred.var Character string giving the names of the predictor variables
-#' of interest. For reasons of computation/interpretation, this should include
-#' no more than three variables.
+#' @param feature_names Character string giving the names of the predictor
+#' variables of interest. For reasons of computation/interpretation, this should
+#' include no more than three variables.
 #'
 #' @param FUN Function used to measure the variability of the partial dependence
 #' values for continuous predictors. If \code{NULL}, the standard deviation is
 #' used (i.e., \code{FUN = sd}). For factors, the range statistic is used (i.e.,
 #' (max - min) / 4).
 #'
-#' @param keep.partial Logical indicating whether or not to return the computed
-#' partial dependence value for the predictor listed in \code{pred.var} in a
-#' separate attribute called \code{"partial"}. Default is \code{FALSE}.
+#' @param keep_partial Logical indicating whether or not to return the computed
+#' partial dependence value for the predictor listed in \code{feature_names} in
+#' a separate attribute called \code{"partial"}. Default is \code{FALSE}.
 #'
 #' @param ... Additional optional arguments to be passed on to
 #' \code{\link[pdp]{partial}}.
 #'
 #' @export
-get_pd_vi_score <- function(object, pred.var, FUN = NULL, keep.partial = FALSE,
-                            ...) {
+get_pd_vi_score <- function(object, feature_names, FUN = NULL,
+                            keep_partial = FALSE, ...) {
+
+  # Variability importance function
   FUN <- if (is.null(FUN)) {  # function to use for continuous predictors
     stats::sd
   } else {
     match.fun(FUN)
   }
-  pd <- pdp::partial(object, pred.var = pred.var, ...)
-  res <- if (is.factor(pd[[pred.var]])) {
+
+  # Compute partial dependence values
+  pd <- pdp::partial(object, pred.var = feature_names, ...)
+
+  # Compute partial dependence-based variable importance scores
+  res <- if (is.factor(pd[[feature_names]])) {
     diff(range(pd[["yhat"]])) / 4
   } else {
     FUN(pd[["yhat"]])
   }
-  if (keep.partial) {
+  if (keep_partial) {
     attr(res, "partial") <- pd
   }
+
+  # Return result
   res
+
 }